Do not add duplicate headers, and clear table for append
Instead of appending the API results non-stop, we clear the table first so that it's fresh data loaded into the table.

In addition, I added a class to the "To" `tr` and jQuery will look for this class. If it doesn't exist then it'll add that `tr`. Otherwise it won't continue to stack it.

Fixes https://github.com/HaschekSolutions/opentrashmail/issues/2